#114 Refactor Popup hierarchy
 Deleted FaultyInputPopup and GameNotFoudPopup.
 WarningPopup now contains all methods.

 <feature, organisation
